requirements.txt
ansicon==1.89.0
asgiref==3.4.1
backports.functools-lru-cache==1.6.1
boto==2.49.0
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==4.0
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.11.2
django-rest-auth==0.9.5
django-taggit==3.1.0
djangorestframework==3.14.0
filelock==3.8.2
gunicorn==20.1.0
idna==2.10
jinxed==1.2.0
jmespath==0.10.0
micropython-ffilib==0.1.2
oauthlib==3.1.0
packaging==21.2
Pillow==9.3.0
platformdirs==2.6.0
psycopg2==2.9.5
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.3
python-editor==1.0.4
python3-openid==3.2.0
pytz==2021.1
readchar==2.0.1
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
tzdata==2021.5
urllib3==1.26.4
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==5.3.0
wincertstore==0.2
